Update: Helper::range() automatically convert id into an integer.

+++ src/Eorm/Eorm.php
@@ -115,7 +115,7 @@ class Eorm
         );
 
         return static::find(
-            Helper::range((int) Server::insertId(static::getServer()), $rows)
+            Helper::range(Server::insertId(static::getServer()), $rows)
         );
     }
